create AMI on the instance
